Polished up the code and fixed some glaring bugs like:
skopos command workflow is alias skopos => $HOME/bin/skopos => sources skopos function. The reason for this is so that every invocation of the command sources the functions. The functions are necessary to export environment into the running interactive shell which cannot be done in subshells or commands running in subshells.

Skopos correctly creates .helm symlink to $HOME/.helm from whichever kraken cluster is being switched to. If a .helm directory exists there, skopos moves it from .helm to dot.helm.

Fixed usage output

Now correctly creates new environment on "create"

Other minor fixes and nits.

OK, so there are still some things that need attention:

Do not rely on the use of "config.yaml" -- find a way to determine arbitrarily named and placed config.yaml

that is it actually as far as I know