Polish up code and fix bugs (#18)
* Some fixes for determining cluster name. Changed main function name to be skopos, which works just fine. I did not think it would, but tests indicate that it works fine. Other minor changes included.

* Removal of jenkinsfile

* Polished up the code and fixed some glaring bugs like:

skopos command workflow is alias skopos => $HOME/bin/skopos => sources skopos function. The reason for this is so that every invocation of the command sources the functions. The functions are necessary to export environment into the running interactive shell which cannot be done in subshells or commands running in subshells.

Skopos correctly creates .helm symlink to $HOME/.helm from whichever kraken cluster is being switched to. If a .helm directory exists there, skopos moves it from .helm to dot.helm.

Fixed usage output

Now correctly creates new environment on "create"

Other minor fixes and nits.

OK, so there are still some things that need attention:

Do not rely on the use of "config.yaml" -- find a way to determine arbitrarily named and placed config.yaml

that is it actually as far as I know

* Fixed usage again.

* Fixed alias work and what skopos script points to.

* Fixed alias work again and what skopos alias points to.

* Edit README for style and readability

* Update for style and lint fixes

* Update for style guide and lint

* Simplify USAGE markdown

* Reduce three heading to one Usage

* Add rubocop configuration

* Update for style and lint fixes stage 1

* Add missing space character

* Made some changes according to a shellcheck

* Ran against shfmt for style. Ran through shellcheck and fixed according to recommendations. Moved init_skopos to skopos.init for better naming convention and more in line with how other tools name similar scripts.

* Run through shfmt and also added fix in install to check $HOME/bin is in $PATH.

* Bugfix to SKOPOS env var failing to properly set because BASH_SCRIPTS env var was not properly setting.

* shfmt on skopos.bash

* Run shfmt over changed files

* Update for shell standard

* Update for two spaces